backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
pymongo==4.9.2
motor==3.6.0
email-validator==2.3.0
transformers==4.47.0
torch==2.5.1
librosa==0.10.2
httpx==0.28.1
groq==0.11.0
